move cloudwatch to separate package
